+++ src/SemanticSearchService.py
@@ -1,5 +1,6 @@
 import openai
 import pandas as pd
+import re
 from openai.embeddings_utils import cosine_similarity
 
 from Util import setup_logger
@@ -198,6 +199,21 @@ class BatchOpenAISemanticSearchService:
         result_df['rank'] = range(1, len(result_df) + 1)
         result_df['docno'] = range(1, len(result_df) + 1)
         return result_df
+
+    @staticmethod
+    def post_process_gpt_input_text_df(gpt_input_text_df, prompt_length_limit):
+        # clean out of prompt texts for existing [1], [2], [3]... in the source_text
+        gpt_input_text_df['text'] = gpt_input_text_df['text'].apply(lambda x: re.sub(r'\[[0-9]+\]', '', x))
+
+        gpt_input_text_df['len_text'] = gpt_input_text_df['text'].apply(lambda x: len(x))
+        gpt_input_text_df['cumsum_len_text'] = gpt_input_text_df['len_text'].cumsum()
+        max_rank = gpt_input_text_df[gpt_input_text_df['cumsum_len_text'] <= prompt_length_limit]['rank'].max() + 1
+        gpt_input_text_df['in_scope'] = gpt_input_text_df['rank'] <= max_rank  # In order to get also the row slightly larger than prompt_length_limit
+        # reorder url_id with url that in scope.
+        url_id_list = gpt_input_text_df['url_id'].unique()
+        url_id_map = dict(zip(url_id_list, range(1, len(url_id_list) + 1)))
+        gpt_input_text_df['url_id'] = gpt_input_text_df['url_id'].map(url_id_map)
+        return gpt_input_text_df
 
 
 if __name__ == '__main__':
