+++ src/gradio_app.py
@@ -1,10 +1,10 @@
 import gradio as gr
 
-from SearchGPTService import SearchGPTService
+from SearchGPT import SearchGPT
 
 
 def query_and_get_answer(search_text):
-    search_gpt_service = SearchGPTService()
+    search_gpt_service = SearchGPT()
     response_text, source_text, data_json = search_gpt_service.query_and_get_answer(search_text)
     return response_text, source_text
